+++ doc/src/releases.rst
@@ -7,6 +7,8 @@
 Changed
 -------
 - XXX Use built-in TLS when sensible.
+- Logs are now handled by the Python logging module. `debug` and `log_file`
+  are not used anymore.
 
 Other
 -----